requirements.txt
wheel
paho-mqtt==2.0.0
pykka==4.0.2
pyserial==3.5
transitions==0.9.0
astral==3.2
RPi.GPIO==0.7.1
adafruit-circuitpython-ads1x15==2.2.25
adafruit-circuitpython-mcp4725==1.4.14
lcdbackpack==1.0.0b1
